groww_pilot/lambdas/stock_fetcher/requirements.txt
# requirements for stock_fetcher

yfinance
pandas
google-generativeai
